backend-app/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# AWS SDK
boto3>=1.35.0
botocore>=1.35.0
requests-aws4auth>=1.3.0

# Graph DB
gremlinpython>=3.7.0

# Vector Search
opensearch-py>=2.7.0

# HTTP Client
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
